Earn money from turbines per generated energy amount

+++ Source/WindEnergyIdle_CPP/DataAssets/ResourceDataAsset.h
@@ -23,8 +23,8 @@ public:
 	EGameResourceType Type;
 
 	UPROPERTY(EditDefaultsOnly)
-	int DefaultAmount;
+	float DefaultAmount;
 	
 	UPROPERTY(EditDefaultsOnly, BlueprintReadOnly)
-	int Amount;
+	float Amount;
 };
